add docstrings for EvaluationRunResult
